feat: splits armor pages into separate elements
