+++ Common/UI/Armor/UIArmorInventoryManager.cs
@@ -3,13 +3,13 @@ using PathOfTerraria.Core.UI;
 namespace PathOfTerraria.Common.UI.Armor;
 
 [Autoload(Side = ModSide.Client)]
-public sealed class UIGearInventoryManager : ModSystem
+public sealed class UIArmorInventoryManager : ModSystem
 {
 	public const string Identifier = $"{PoTMod.ModName}:Inventory";
 
 	public override void OnWorldLoad()
 	{
-		UIManager.TryEnableOrRegister(Identifier, "Vanilla: Inventory", new UIGearInventory(), 1);
+		UIManager.TryEnableOrRegister(Identifier, "Vanilla: Inventory", new UIArmorInventory(), 1);
 
 		// Re-initialize the state - redundant on first load, may want to change in the future
 		UIManager.TryRefresh(Identifier);
